+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.13-unstable.22"></a>
+## [0.2.13-unstable.22](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.21...aws-amplify-react-native@0.2.13-unstable.22) (2018-06-11)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.13-unstable.21"></a>
+## [0.2.13-unstable.21](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.20...aws-amplify-react-native@0.2.13-unstable.21) (2018-06-08)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.13-unstable.20"></a>
+## [0.2.13-unstable.20](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.19...aws-amplify-react-native@0.2.13-unstable.20) (2018-06-08)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.13-unstable.19"></a>
+## [0.2.13-unstable.19](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.18...aws-amplify-react-native@0.2.13-unstable.19) (2018-06-07)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.13-unstable.18"></a>
+## [0.2.13-unstable.18](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.17...aws-amplify-react-native@0.2.13-unstable.18) (2018-06-06)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.13-unstable.17"></a>
+## [0.2.13-unstable.17](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.16...aws-amplify-react-native@0.2.13-unstable.17) (2018-06-05)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="0.2.13-unstable.16"></a>
 ## [0.2.13-unstable.16](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.15...aws-amplify-react-native@0.2.13-unstable.16) (2018-06-04)